+++ src/WAVTrack.cpp
@@ -28,7 +28,7 @@ void WAVTrack::analyze_beatgrid() {
     // should print "  → Estimated beats: <beats>  → Precision factor: 1.0 (uncompressed audio)"
     std::cout << "[WAVTrack::analyze_beatgrid] Analyzing beat grid for: \"" << title << "\"\n";
     double beats = (duration_seconds / 60.0) * bpm;
-    std::cout << "  → Estimated beats: " << beats << "  → Precision factor: 1 (uncompressed audio)" << std::endl;
+    std::cout << "  → Estimated beats: " << (int)beats << "  → Precision factor: 1 (uncompressed audio)" << std::endl;
 }
 
 double WAVTrack::get_quality_score() const {
